\name{fgpt-package}
\alias{fgpt-package}
\docType{package}
\title{
Floating Grid Permutation Technique
}
\description{
FGPT is a spatially restricted permutation technique. The package contains both low level functions to produce permuted datasets as well as higher level functions to perform spatially restricted permutation tests and functions to summarize and plot the results.
}
\details{
\code{fgeasy} is the single step version of the FGPT, which is most suitable for first time users. Please read the vignette for a more extensive discription of the method and functions.
}
\author{
Reinder Radersma & Ben Sheldon
}
\references{
Reinder Radersma & Ben C. Sheldon, 2015. A new permutation technique to explore and control for spatial autocorrelation. Submitted
}
\keyword{ package }
