\name{Gpulex}
\docType{data}
\alias{Gpulex}
\alias{Gp.xy}
\alias{f.pheno}
\alias{m.pheno}
\alias{m.pheno2}
\title{Size assortative pairing in a freshwater shrimp (\emph{Gammarus pulex})}
\description{
This is a simulated data set of geographical locations and body size measurements for two populations of fresh water Crustaceans. Both populations share the geographical locations and female phenotypes, but not the male phenotypes (\code{m.pheno} and \code{m.pheno2} respectively). The first population does not show size assortative pairing, while the second does.
}
\usage{Gpulex}
\format{An R data object containing one two-column matrix with geographical coordinates (\code{Gp.xy}) and three vectors with female (\code{f.pheno})  and male body size measurements (\code{m.pheno} and \code{m.pheno2}).}
\source{Code to produce this data set can be found in electronic supplementary materials of the following reference.}
\references{
Reinder Radersma & Ben C. Sheldon, 2014. A new permutation test to explore and control for spatial autocorrelation in point pattern data. Manuscript submitted to MEE
}
\keyword{datasets}