\name{Pmajor}
\docType{data}
\alias{Pmajor}
\title{Inbreeding avoidance in the Great Tit (\emph{Parus major})}
\description{
This is a simulated data set of geographical locations and relatedness measures for two populations of Great Tits, a small passerine. Both populations share the geographical locations, but not the relatedness tables. The first population does not show inbreeding avoidance, while the second does.
}
\usage{Pmajor}
\format{An R data object containing one two-column matrix with geographical coordinates (\code{xy}) and two relatedness tables between all females and males of the populations (\code{rel1} and \code{rel2} respectively).}
\source{Code to produce this data set can be found in electronic supplementary materials of the following reference.}
\references{
Reinder Radersma & Ben C. Sheldon, 2014. A new permutation test to explore and control for spatial autocorrelation in point pattern data. Manuscript submitted to MEE
}
\keyword{datasets}
